Json CRUD fixed

+++ src/components/JsonBlock.js
@@ -1,19 +1,24 @@
 import React from 'react';
 import JsonItem from './JsonItem';
 
-var JsonBlock = ({ item, handleItemChange, deleteItem }) => {
+var JsonBlock = ({ item, handleItemChange, deleteItem, addSubItem }) => {
     // console.log(onDelete)
-    if(item.type === "array" || item.type === "object"){
+    if((item.type === "array" || item.type === "object") && item.value.length !== 0){
         return (
             <div className="item-holder">
-                <JsonItem item={item} handleItemChange={handleItemChange} deleteItem={deleteItem} />
+                <JsonItem 
+                    item={item} 
+                    handleItemChange={handleItemChange} 
+                    deleteItem={deleteItem} 
+                    addSubItem={addSubItem} />
 
                 {item.value.map(item => {
-                        return <JsonBlock 
+                    return <JsonBlock 
                         key={item.id} 
                         item={item}
                         handleItemChange={handleItemChange}
                         deleteItem={deleteItem}
+                        addSubItem={addSubItem}
                     />
                 })}
             </div>
@@ -21,7 +26,12 @@ var JsonBlock = ({ item, handleItemChange, deleteItem }) => {
     } else {
         return (
               <div className="item-holder">
-                <JsonItem key={item.id} item={item} handleItemChange={handleItemChange} deleteItem={deleteItem} />
+                <JsonItem 
+                key={item.id} 
+                item={item} 
+                handleItemChange={handleItemChange} 
+                deleteItem={deleteItem}
+                addSubItem={addSubItem} />
             </div>
         )
     }
